BUG: Fix typos in the acknowledgement and help texts.
See https://projectreporter.nih.gov/project_info_details.cfm?aid=9125815&icde=35404927
for justification of the correction I made to the acknowledgement text.

The help text was reworded for reasons of English grammar only. It may
need to be reworded at some point as it may be overly specific and give
potential users the impression that the module cannot hel them for their
use case.

All text was reflowed to wrap at 72 columns or fewer. Whitespace in
these values appears to be ignored by Slicer.